requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
Pillow>=8.0.0
nltk>=3.6.0
rouge>=1.0.1
tqdm>=4.62.0
matplotlib>=3.4.0
tensorboard>=2.6.0
transformers>=4.5.0